add db for multiple users

+++ dbhelper.py
@@ -6,29 +6,28 @@ class DBHelper:
 		self.conn = sqlite3.connect(dbname)
 
 	def setup(self):
-		stmt = "CREATE TABLE IF NOT EXISTS items (description text)"
-		self.conn.execute(stmt)
-		self.conn.commit()
+		tblstmt = "CREATE TABLE IF NOT EXISTS items (description text, owner text)"
+        itemidx = "CREATE INDEX IF NOT EXISTS itemIndex ON items (description ASC)" 
+        ownidx = "CREATE INDEX IF NOT EXISTS ownIndex ON items (owner ASC)"
+        self.conn.execute(tblstmt)
+        self.conn.execute(itemidx)
+        self.conn.execute(ownidx)
+        self.conn.commit()
 
-	def add_item(self, item_text):
-		stmt = "INSERT INTO items (description) VALUES (?)"
-		args = (item_text, )
-		self.conn.execute(stmt, args)
-		self.conn.commit()
+    def add_item(self, item_text, owner):
+        stmt = "INSERT INTO items (description, owner) VALUES (?, ?)"
+        args = (item_text, owner)
+        self.conn.execute(stmt, args)
+        self.conn.commit()
 
-	def delete_item(self, item_text):
-		stmt = "INSERT INTO items (description) VALUES (?)"
-		args = (item_text, )
-		self.conn.execute(stmt ,args)
-		self.conn.commit()
+    def delete_item(self, item_text, owner):
+        stmt = "DELETE FROM items WHERE description = (?) AND owner = (?)"
+        args = (item_text, owner )
+        self.conn.execute(stmt, args)
+        self.conn.commit()
 
-	def delete_item(self, item_text):
-		stmt = "DELETE FROM items WHERE description = (?)"
-		args = (item_text, )
-		self.conn.execute(stmt, args)
-		self.conn.commit()
-
-	def get_items(self):
-		stmt = "SELECT description FROM items"
-		return [x[0] for x in self.conn.execute(stmt)]
+    def get_items(self, owner):
+        stmt = "SELECT description FROM items WHERE owner = (?)"
+        args = (owner, )
+        return [x[0] for x in self.conn.execute(stmt, args)]
